[SIM] Relax trailer verification for multi-image
The simulator does trailer verification to ensure the status is what
would be expected when operating a swap. When doing a multi-image swap
upgrade one of the images must be in a known status but others might not
be, depending on which image number is being upgraded at the moment.
This patch relaxes the verification to assure at least one is in a known
status.

Signed-off-by: Fabio Utzig <utzig@apache.org>