WIP: Work on enabling -Wconvert in boot code
This enables `-Wconversion` and `-Wno-sign-conversion` options to the build
of the bootutil code.  The rest of the changes are enough casts to get
everything to compile.

Casting is not usually the correct solution, and what needs to happen,
instead, is to propagate the correct types to various places that these
values are used.  This change is instead here to show us what values
need these changes.

Signed-off-by: David Brown <david.brown@linaro.org>